made a component a client component
